third_party/WebKit/LayoutTests/virtual/implsidepainting/inspector/tracing/paint-command-log-nodes-expected.txt

Tests snapshot command log for trace-based Timeline paint event


Paint command log:
    div#a drawRect
    div#b drawRect
    div#alttext-container drawRect
    div#alttext-container drawRect
    div#alttext-container drawRect
    div#alttext-container drawRect
    img#alttext-image drawBitmapRectToRect
    rect#e drawRect
